fixed scss format in _sass file, now how to preprocess correctly?
_Scss and minima folders are now properly configurated, and I have done a few edits.

However I am having a bit of an issue with the preprocessor, and figuring out how to store the .css. Other than that I need to test the OG I input into the files, and also configure with github. However, I think this site should be almost done now, finally